This commit tests interpolating values across the triangle face.
There are a lot of changes here, but the gist of it is:
In the tutorial, the instructor gives up on the triangle sweep to
opt for a parametric model that evaluates the barycentric
coordinates of each pixel in the triangle's bounding box.
In my extremely limited testing, this method is 250 times slower
than the triangle line-sweep.

My thought was that I can just calculate the barycentric coords
for the left and right -most pixels for the line-sweep algorithm
and then lerp across the face of the triangle.

So this codebase includes code for doing both methods.
I've rendered pictures and they look the same to my eyes, but I
have not done extensive testing.  Currently in the process of
calculating image diffs.